+++ src/wtforms/fields/core.py
@@ -89,8 +89,8 @@ class Field:
             If provided, a dictionary which provides default keywords that
             will be given to the widget at render time.
         :param name:
-            The HTML name of this field. The default value is the python
-            field name.
+            The HTML name of this field. The default value is the Python
+            attribute name.
         :param _form:
             The form holding this field. It is passed by the form itself during
             construction. You should never pass this value yourself.